Change sequence number type to int64 offchain and uint64 onchain (#206)
* chore: change sequence number type on chain to uint64 and offchain to int64

- swap ordering within CCIPMessage to reduce the number of slots in the struct
- we use int64 offchain due to the postgres datatypes

+++ contracts/src/v0.8/ccip/applications/SenderDapp.sol
@@ -43,7 +43,7 @@ contract SenderDapp is TypeAndVersionInterface {
     IERC20[] memory tokens,
     uint256[] memory amounts,
     address executor
-  ) external returns (uint256 sequenceNumber) {
+  ) external returns (uint64 sequenceNumber) {
     if (destinationAddress == address(0)) revert InvalidDestinationAddress(destinationAddress);
     bytes memory options;
     address originalSender = msg.sender;
